mqttmpd: allow to a series of playlists to iterate through
This commit extends playlist/set:
When multiple playlists are provided (space-seperated),
then mqttmpd will iterate through them, i.e.
it will remember the last one, and select the next one.
In case mqttmpd needs to pick a playlist, it is reported via
playlist/selected.

This helps to implement this scenario:
* Each time I press a button in a relative short period of time,
  I will select the next playlist in my list (assigned to that button).
* For feedback, the selected playlist is reported back.
* If I wait longer, pressing that button will stop mpd.
  In that case, an empty topic is reported.